Move to CBOR trait from configured formats
This commit refactors the trait implementation to be specific to the
CBOR wire format instead of taking a collection of possible wire
formats, simplifying the process of resolving the protocol to use
for generated code.

It also moves the package to `smithy-protocol-traits`, removing the
pluralization on `protocols`.
