+++ settings.gradle
@@ -33,4 +33,4 @@ include ":smithy-rules-engine"
 include ":smithy-smoke-test-traits"
 include ":smithy-syntax"
 include ":smithy-aws-endpoints"
-include ":smithy-protocols-traits"
+include ":smithy-protocol-traits"
