+++ smithy-protocol-traits/src/main/java/software/amazon/smithy/protocol/traits/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/**
+ * Defines protocols for Smithy.
+ */
+@SmithyUnstableApi
+package software.amazon.smithy.protocol.traits;
+
+import software.amazon.smithy.utils.SmithyUnstableApi;
