+++ smithy-protocol-traits/src/main/resources/META-INF/services/software.amazon.smithy.model.traits.TraitService
@@ -1,0 +1,1 @@
+software.amazon.smithy.protocol.traits.Rpcv2CborTrait$Provider
