+++ smithy-protocol-traits/src/main/resources/META-INF/smithy/manifest
