--- smithy-protocols-traits/src/main/java/software/amazon/smithy/protocols/traits/package-info.java
@@ -1,22 +1,0 @@
-/*
- * Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License").
- * You may not use this file except in compliance with the License.
- * A copy of the License is located at
- *
- *  http://aws.amazon.com/apache2.0
- *
- * or in the "license" file accompanying this file. This file is distributed
- * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing
- * permissions and limitations under the License.
- */
-
-/**
- * Defines protocols for Smithy.
- */
-@SmithyUnstableApi
-package software.amazon.smithy.protocols.traits;
-
-import software.amazon.smithy.utils.SmithyUnstableApi;
