--- smithy-protocols-traits/src/main/resources/META-INF/services/software.amazon.smithy.model.validation.Validator
@@ -1,1 +1,0 @@
-software.amazon.smithy.protocols.traits.Rpcv2TraitValidator
